Move various set_px_cb_xx functions to lv_hal_disp.c
