+++ src/misc/lv_snapshot.c
@@ -22,7 +22,6 @@
 /**********************
  *  STATIC PROTOTYPES
  **********************/
-static void set_set_px_cb(lv_disp_drv_t * disp_drv, lv_img_cf_t cf);
 
 /**********************
  *  STATIC VARIABLES
@@ -95,7 +94,7 @@ lv_img_dsc_t * lv_snapshot_take(lv_obj_t * obj, lv_img_cf_t cf)
     driver.draw_buf = &draw_buf;
     driver.hor_res = lv_disp_get_hor_res(disp_old);
     driver.ver_res = lv_disp_get_ver_res(disp_old);
-    set_set_px_cb(&driver, cf);
+    lv_disp_drv_use_generic_set_px_cb(&driver, cf);
 
     disp = lv_disp_drv_register(&driver);
     if(disp == NULL) {
@@ -158,158 +157,3 @@ void lv_snapshot_free(lv_img_dsc_t * dsc)
 /**********************
  *   STATIC FUNCTIONS
  **********************/
-
-static void set_px_true_color_alpha(lv_disp_drv_t * disp_drv, uint8_t * buf,
-                                    lv_coord_t buf_w, lv_coord_t x,
-                                    lv_coord_t y, lv_color_t color,
-                                    lv_opa_t opa);
-
-static void set_px_cb_alpha1(lv_disp_drv_t * disp_drv, uint8_t * buf,
-                             lv_coord_t buf_w, lv_coord_t x, lv_coord_t y,
-                             lv_color_t color, lv_opa_t opa);
-
-static void set_px_cb_alpha2(lv_disp_drv_t * disp_drv, uint8_t * buf,
-                             lv_coord_t buf_w, lv_coord_t x, lv_coord_t y,
-                             lv_color_t color, lv_opa_t opa);
-
-static void set_px_cb_alpha4(lv_disp_drv_t * disp_drv, uint8_t * buf,
-                             lv_coord_t buf_w, lv_coord_t x, lv_coord_t y,
-                             lv_color_t color, lv_opa_t opa);
-
-static void set_px_cb_alpha8(lv_disp_drv_t * disp_drv, uint8_t * buf,
-                             lv_coord_t buf_w, lv_coord_t x, lv_coord_t y,
-                             lv_color_t color, lv_opa_t opa);
-
-static void set_px_alpha_generic(lv_img_dsc_t * d, lv_coord_t x, lv_coord_t y,
-                                 lv_color_t color, lv_opa_t opa);
-
-static void set_px_cb_alpha1(lv_disp_drv_t * disp_drv, uint8_t * buf,
-                             lv_coord_t buf_w, lv_coord_t x, lv_coord_t y,
-                             lv_color_t color, lv_opa_t opa)
-{
-    (void)disp_drv; /*Unused*/
-
-    if(opa <= LV_OPA_MIN)
-        return;
-    lv_img_dsc_t d;
-    d.data = buf;
-    d.header.w = buf_w;
-    d.header.cf = LV_IMG_CF_ALPHA_1BIT;
-
-    set_px_alpha_generic(&d, x, y, color, opa);
-}
-
-static void set_px_cb_alpha2(lv_disp_drv_t * disp_drv, uint8_t * buf,
-                             lv_coord_t buf_w, lv_coord_t x, lv_coord_t y,
-                             lv_color_t color, lv_opa_t opa)
-{
-    (void)disp_drv; /*Unused*/
-
-    if(opa <= LV_OPA_MIN)
-        return;
-    lv_img_dsc_t d;
-    d.data = buf;
-    d.header.w = buf_w;
-    d.header.cf = LV_IMG_CF_ALPHA_2BIT;
-
-    set_px_alpha_generic(&d, x, y, color, opa);
-}
-
-static void set_px_cb_alpha4(lv_disp_drv_t * disp_drv, uint8_t * buf,
-                             lv_coord_t buf_w, lv_coord_t x, lv_coord_t y,
-                             lv_color_t color, lv_opa_t opa)
-{
-    (void)disp_drv; /*Unused*/
-
-    if(opa <= LV_OPA_MIN)
-        return;
-    lv_img_dsc_t d;
-    d.data = buf;
-    d.header.w = buf_w;
-    d.header.cf = LV_IMG_CF_ALPHA_4BIT;
-
-    set_px_alpha_generic(&d, x, y, color, opa);
-}
-
-static void set_px_cb_alpha8(lv_disp_drv_t * disp_drv, uint8_t * buf,
-                             lv_coord_t buf_w, lv_coord_t x, lv_coord_t y,
-                             lv_color_t color, lv_opa_t opa)
-{
-    (void)disp_drv; /*Unused*/
-
-    if(opa <= LV_OPA_MIN)
-        return;
-    lv_img_dsc_t d;
-    d.data = buf;
-    d.header.w = buf_w;
-    d.header.cf = LV_IMG_CF_ALPHA_8BIT;
-
-    set_px_alpha_generic(&d, x, y, color, opa);
-}
-
-static void set_px_alpha_generic(lv_img_dsc_t * d, lv_coord_t x, lv_coord_t y,
-                                 lv_color_t color, lv_opa_t opa)
-{
-    d->header.always_zero = 0;
-    d->header.h = 1; /*Doesn't matter*/
-
-    uint8_t br = lv_color_brightness(color);
-    if(opa < LV_OPA_MAX) {
-        uint8_t bg = lv_img_buf_get_px_alpha(d, x, y);
-        br = (uint16_t)((uint16_t)br * opa + (bg * (255 - opa))) >> 8;
-    }
-
-    lv_img_buf_set_px_alpha(d, x, y, br);
-}
-
-static void set_px_true_color_alpha(lv_disp_drv_t * disp_drv, uint8_t * buf,
-                                    lv_coord_t buf_w, lv_coord_t x,
-                                    lv_coord_t y, lv_color_t color,
-                                    lv_opa_t opa)
-{
-    (void)disp_drv; /*Unused*/
-
-    if(opa <= LV_OPA_MIN)
-        return;
-    lv_img_dsc_t d;
-    d.data = buf;
-    d.header.always_zero = 0;
-    d.header.h = 1; /*Doesn't matter*/
-    ;
-    d.header.w = buf_w;
-    d.header.cf = LV_IMG_CF_TRUE_COLOR_ALPHA;
-
-    lv_color_t bg_color = lv_img_buf_get_px_color(&d, x, y, lv_color_black());
-    lv_opa_t bg_opa = lv_img_buf_get_px_alpha(&d, x, y);
-
-    lv_opa_t res_opa;
-    lv_color_t res_color;
-
-    lv_color_mix_with_alpha(bg_color, bg_opa, color, opa, &res_color, &res_opa);
-
-    lv_img_buf_set_px_alpha(&d, x, y, res_opa);
-    lv_img_buf_set_px_color(&d, x, y, res_color);
-}
-
-static void set_set_px_cb(lv_disp_drv_t * disp_drv, lv_img_cf_t cf)
-{
-    switch(cf) {
-        case LV_IMG_CF_TRUE_COLOR_ALPHA:
-            disp_drv->set_px_cb = set_px_true_color_alpha;
-            break;
-        case LV_IMG_CF_ALPHA_1BIT:
-            disp_drv->set_px_cb = set_px_cb_alpha1;
-            break;
-        case LV_IMG_CF_ALPHA_2BIT:
-            disp_drv->set_px_cb = set_px_cb_alpha2;
-            break;
-        case LV_IMG_CF_ALPHA_4BIT:
-            disp_drv->set_px_cb = set_px_cb_alpha4;
-            break;
-        case LV_IMG_CF_ALPHA_8BIT:
-            disp_drv->set_px_cb = set_px_cb_alpha8;
-            break;
-        default:
-            disp_drv->set_px_cb = NULL;
-    }
-}
